LPS-71628 Revert "LPS-71628 Update impl to take the parametrized strategy into account"
This reverts commit 6000de0d01f9cb055adf23c908022ad345f5027a.
