+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-io/src/main/java/com/liferay/dynamic/data/mapping/io/DDMFormLayoutJSONDeserializer.java
@@ -21,8 +21,6 @@ import com.liferay.dynamic.data.mapping.model.DDMFormLayoutColumn;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONObject;
 
-import java.util.function.Function;
-
 /**
  * @author Marcellus Tavares
  */
@@ -34,8 +32,15 @@ public interface DDMFormLayoutJSONDeserializer {
 
 	public DDMFormLayout deserialize(
 			String serializedDDMFormLayout,
-			Function<JSONObject, DDMFormLayoutColumn>
+			DDMFormLayoutColumnJSONObjectTransformer
 				ddmFormLayoutColumnJSONObjectTransformer)
 		throws PortalException;
+
+	@FunctionalInterface
+	public interface DDMFormLayoutColumnJSONObjectTransformer {
+
+		public DDMFormLayoutColumn transform(JSONObject jsonObject);
+
+	}
 
 }
